Minor update
Add throw exception if resource is not found

+++ src/main/java/com/spring/rest/customer_rest/service/CustomerServiceImpl.java
@@ -4,6 +4,7 @@ import com.spring.rest.customer_rest.dao.CustomerDAO;
 import com.spring.rest.customer_rest.dto.CustomerDTO;
 import com.spring.rest.customer_rest.entity.Customer;
 import com.spring.rest.customer_rest.exceptions.EmailAlreadyExistsException;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,6 +23,9 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional
     public CustomerDTO getCustomerById(int id) {
         Customer customer = customerDAO.getCustomerById(id);
+        if(customer == null) {
+            throw new EntityNotFoundException();
+        }
         return customerDTOMapper.entityToDTO(customer);
     }
 
@@ -29,6 +33,9 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional
     public List<CustomerDTO> getAllCustomers() {
         List<Customer> customers = customerDAO.getAllCustomers();
+        if(customers.isEmpty()) {
+            throw new EntityNotFoundException();
+        }
         return customers.stream().map(customerDTOMapper::entityToDTO).toList();
     }
 
@@ -46,6 +53,9 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     @Transactional
     public void deleteCustomerById(int id) {
+        if(customerDAO.getCustomerById(id) == null) {
+            throw new EntityNotFoundException();
+        }
         customerDAO.deleteCustomerById(id);
     }
 
@@ -53,6 +63,9 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional
     public CustomerDTO updateCustomer(int id, CustomerDTO customerDTO) {
         Customer customer = customerDAO.getCustomerById(id);
+        if(customer == null) {
+            throw new EntityNotFoundException();
+        }
         Customer customerFromDTO = customerDTOMapper.dtoToEntity(customerDTO);
 
         customer.setFullName(customerFromDTO.getFullName());
